#208 refactor: Category가 BaseEntity를 상속하도록 변경, 그에 따라 bookCategory는 ACTIVE만 조회하도록 수정
