+++ src/main/java/com/floney/floney/book/entity/Category.java
@@ -1,37 +1,28 @@
 package com.floney.floney.book.entity;
 
+import com.floney.floney.common.entity.BaseEntity;
 import java.util.List;
 import javax.persistence.CascadeType;
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.FetchType;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
 import javax.persistence.Inheritance;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
 import javax.persistence.OneToMany;
 import lombok.AccessLevel;
-import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.experimental.SuperBuilder;
 import org.hibernate.annotations.OnDelete;
 import org.hibernate.annotations.OnDeleteAction;
 
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@EqualsAndHashCode
 @Getter
 @Entity
 @Inheritance
 @SuperBuilder
-public abstract class Category {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(nullable = false)
-    private Long id;
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public abstract class Category extends BaseEntity {
 
     @Column(length = 10)
     private String name;
@@ -44,7 +35,7 @@ public abstract class Category {
     @OnDelete(action = OnDeleteAction.CASCADE)
     private Category parent;
 
-    public Category(String name, Category parent) {
+    protected Category(String name, Category parent) {
         this.name = name;
         this.parent = parent;
     }
